refactor: вынести телеграм-бот в отдельный пакет
