+++ docs/PAYMENTS.md
@@ -19,3 +19,11 @@
 1. Используйте `/status`, чтобы убедиться в отсутствии активной подписки.
 2. Свяжитесь с поддержкой и предоставьте данные платежа.
 3. При необходимости администратор отменит подписку и выполнит возврат.
+
+## Структура кода бота
+Основная логика вынесена в пакет `pipeline.telegram_bot`:
+- `menus.py` — стартовые и платёжные меню;
+- `affiliates.py` — партнёрская программа;
+- `admin.py` — административные функции.
+Файл `trading/payments_bot.py` собирает приложение и регистрирует хэндлеры.
+
